Completed Pastitsio page of website.
In this update I added the ingredients list as an unordered list to the website under a heading3. I added all of the steps of the recipe as an ordered list underneath.

In the next update i plan to add a similar page in structure of a different recipe for tiropita. I will reuse this template from the pastitsio page.