Added second Page for Tiropita Recipe.
I copied the template that I had made in creating the Pastitsio recipe page. I wrote the changes in recipes and ingredients to fit that of the recipe for the tiropita.

I also made an ordered list of the two pages on the websit. I plan to make links on both pages to go back to the home page.