Reorganized Website with divs and added homepage links
I went back through and organized the main parts of my code into divs and restructured the recipe pages to be within <main> instead of structured as whole web pages on their own. I also added a return to the top link on the homepage as well as homepage links on the recipe pages.